[Shaman] Implement an opt-in deck-based RNG for Deeply Rooted Elements
This commit introduces a customized "deck-of-cards" style rng process to
Deeply Rooted Elements proc chance. The deck draws per consumed resource
(i.e., Maelstrom, or Maelstrom Weapon stack). The randomization of the
deck is implemented such that a single consumption event (a spell cast)
cannot proc DRE more than once.

By default, the deck size is calculated based on the given spell data
proc chance and the experimentally arrived value of 2 successful draws
per deck. This gets us a deck of 2/333 for Enhancement, and 2/1723 for
Elemental.

The new algorithm can be enabled through the new player-scope option
`shaman.use_new_dre` (default 0). Additionally, the following options
can be used to control the deck behavior:

- `shaman.dre_draw_success` (default 2, scope player) -- defines the
  number of successful draws in the deck
- `shaman.n_dre_draws` (default 333 or 1723, scope player) -- defines
  the total number of cards in the deck

Note that if the algorithm is unable to find a shuffled solution for the
deck (i.e., all the cards satisfy the condition that they are at least
"max draws" apart), the simulator will cancel the iteration.